# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de produtos",
        "description": "API para acessar os recursos de produtos",
        "contact": {
            "name": "Suporte da minha API",
            "email": "gian15249@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza o login do usuário na aplicação",
                "description": "Retorna o token de acesso do usuário para autenticação nas rotas protegidas da API",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "gian.ramalho@teste.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "authorization": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LjAuMC4xOjgwMDAvYXBpL2xvZ2luIiwiaWF0IjoxNjc5OTU5MDUxLCJleHAiOjE2Nzk5NjI2NTEsIm5iZiI6MTY3OTk1OTA1MSwianRpIjoiUUwzUnJqaXBpOGcyY0dNSyIsInN1YiI6IjEiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.7YnEcK0jC1TkzSnf5whzDgTtafJP3Zqx0hdebBpxDOc"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Logout user",
                "description": "Remove o token de acesso do usuário para autenticação nas rotas protegidas da API",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logout feito com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Desconectado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/produtos/{id}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Exibe informações de um produto específico",
                "description": "Retorna as informações de um produto, baseado no ID informado.",
                "operationId": "showProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do produto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Apple iPhone 11 (64 GB) Preto"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2951.99
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Tela LCD Liquid Retina HD de 6,1 polegadas; Resistente a água e pó (até 30 minutos à profundidade máxima de 2 metros, IP68); Sistema de câmara dupla de 12 MP (Ultra grande angular, Grande angular e Teleobjetiva); modo Noite, modo Retrato e vídeos em 4K a 60 fps; Sistema de câmara frontal de 12 MP com modo Retrato, vídeos em 4K e câmara lenta; Face ID para autenticação segura; Conteúdo da caixa iPhone com iOS 14, cabo USB-C para Lightning, documentação"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-03-28T00:39:46.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-03-28T00:39:46.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nenhum produto encontrado com o ID especificado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}